+++ source/dotnet/Services/WebApi.IntegrationTests/WebApi/DatabaseHealthCheckTests.cs
@@ -23,33 +23,14 @@ using Xunit.Abstractions;
 
 namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi;
 
-[Collection(nameof(WholesaleWebApiCollectionFixture))]
-public class DatabaseHealthCheckTests :
-    WebApiTestBase<WholesaleWebApiFixture>,
-    IClassFixture<WholesaleWebApiFixture>,
-    IClassFixture<WebApiFactory>,
-    IAsyncLifetime
+public class DatabaseHealthCheckTests : WebApiTestBase
 {
-    private readonly HttpClient _client;
-
     public DatabaseHealthCheckTests(
         WholesaleWebApiFixture wholesaleWebApiFixture,
         WebApiFactory factory,
         ITestOutputHelper testOutputHelper)
-        : base(wholesaleWebApiFixture, testOutputHelper)
+        : base(wholesaleWebApiFixture, factory, testOutputHelper)
     {
-        _client = factory.CreateClient();
-    }
-
-    public Task InitializeAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public Task DisposeAsync()
-    {
-        _client.Dispose();
-        return Task.CompletedTask;
     }
 
     [Fact]
@@ -59,7 +40,7 @@ public class DatabaseHealthCheckTests :
         await Fixture.DatabaseManager.DeleteDatabaseAsync();
 
         // Act
-        var actualResponse = await _client.GetAsync(HealthChecksConstants.ReadyHealthCheckEndpointRoute);
+        var actualResponse = await Client.GetAsync(HealthChecksConstants.ReadyHealthCheckEndpointRoute);
 
         // Assert
         actualResponse.StatusCode.Should().Be(HttpStatusCode.ServiceUnavailable);
