+++ source/dotnet/Services/WebApi.IntegrationTests/WebApi/V2/OpenApiSpecificationTests.cs
@@ -15,49 +15,38 @@
 using System.Net;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.TestCommon.Fixture.WebApi;
 using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.Fixtures.WebApi;
+using Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi.V3;
 using FluentAssertions;
+using Newtonsoft.Json.Linq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi;
+namespace Energinet.DataHub.Wholesale.WebApi.IntegrationTests.WebApi.V2;
 
-[Collection(nameof(WholesaleWebApiCollectionFixture))]
-public class SwaggerTests :
-    WebApiTestBase<WholesaleWebApiFixture>,
-    IClassFixture<WholesaleWebApiFixture>,
-    IClassFixture<WebApiFactory>,
-    IAsyncLifetime
+public class OpenApiSpecificationTests : WebApiTestBase
 {
-    private readonly HttpClient _client;
-
-    public SwaggerTests(
+    public OpenApiSpecificationTests(
         WholesaleWebApiFixture wholesaleWebApiFixture,
         WebApiFactory factory,
         ITestOutputHelper testOutputHelper)
-        : base(wholesaleWebApiFixture, testOutputHelper)
+        : base(wholesaleWebApiFixture, factory, testOutputHelper)
     {
-        _client = factory.CreateClient();
-    }
-
-    public Task InitializeAsync()
-    {
-        return Task.CompletedTask;
-    }
-
-    public Task DisposeAsync()
-    {
-        _client.Dispose();
-        return Task.CompletedTask;
     }
 
     [Fact]
-    public async Task Given_WebAPI_When_GettingSwaggerUi_Then_ReturnsHtmlPage()
+    public async Task Given_WebAPI_When_GettingOpenApiSpec_Then_ReturnsSpecificationInJson()
     {
+        // Arrange
+        const string expectedOpenApiSpecUrl = "/swagger/v2/swagger.json";
+
         // Act
-        var actualResponse = await _client.GetAsync("/swagger/index.html");
+        var actualResponse = await Client.GetAsync(expectedOpenApiSpecUrl);
 
         // Assert
-        actualResponse.Content.Headers.ContentType!.ToString().Should().StartWith("text/html");
         actualResponse.StatusCode.Should().Be(HttpStatusCode.OK);
+
+        var actualContent = await actualResponse.Content.ReadAsStringAsync();
+        var actualContentJObject = JObject.Parse(actualContent);
+        actualContentJObject.Should().NotBeNull();
     }
 }
